feat: add secret to data exporter (#736)
